requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.9
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
asgiref==3.8.1
astroid==3.3.5
attrs==24.2.0
autopep8==2.0.4
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
blis==1.0.1
build==1.2.2
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.11
click==8.1.7
cloudpathlib==0.19.0
coloredlogs==15.0.1
confection==0.1.5
cryptography==43.0.1
cymem==2.0.8
dataclasses-json==0.6.7
deepdiff==8.0.1
Deprecated==1.2.14
dill==0.3.9
distro==1.9.0
docstring-to-markdown==0.15
durationpy==0.9
emoji==2.14.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.115.0
filelock==3.16.1
filetype==1.2.0
flake8==7.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.9.0
google-auth==2.35.0
googleapis-common-protos==1.65.0
greenlet==3.1.1
grpcio==1.66.2
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.1
humanfriendly==10.0
idna==3.10
importlib_metadata==8.4.0
importlib_resources==6.4.5
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kubernetes==31.0.0
langchain==0.3.2
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-core==0.3.9
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langcodes==3.4.1
langdetect==1.0.9
langsmith==0.1.131
language_data==1.2.0
lxml==5.3.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mccabe==0.7.0
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
oauthlib==3.2.2
olefile==0.47
onnxruntime==1.19.2
openai==1.51.0
opentelemetry-api==1.27.0
opentelemetry-exporter-otlp-proto-common==1.27.0
opentelemetry-exporter-otlp-proto-grpc==1.27.0
opentelemetry-instrumentation==0.48b0
opentelemetry-instrumentation-asgi==0.48b0
opentelemetry-instrumentation-fastapi==0.48b0
opentelemetry-proto==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-semantic-conventions==0.48b0
opentelemetry-util-http==0.48b0
orderly-set==5.2.2
orjson==3.10.7
overrides==7.7.0
packaging==24.1
parso==0.8.4
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
posthog==3.7.0
preshed==3.0.9
protobuf==4.25.5
psutil==6.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pydocstyle==6.3.0
pyflakes==3.2.0
Pygments==2.18.0
pylint==3.3.1
pypdf==5.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2024.4.27
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
python-magic==0.4.27
python-multipart==0.0.12
python-oxmsg==0.0.1
pytoolconfig==1.3.1
PyYAML==6.0.2
RapidFuzz==3.10.0
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.2
rope==1.13.0
rsa==4.9
shellingham==1.5.4
six==1.16.0
smart-open==7.0.5
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.6
spacy==3.8.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.35
srsly==2.4.8
starlette==0.38.6
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
thinc==8.3.2
tiktoken==0.8.0
tokenizers==0.20.0
tomli==2.0.2
tomlkit==0.13.2
tqdm==4.66.5
typer==0.12.5
typing-inspect==0.9.0
typing_extensions==4.12.2
ujson==5.10.0
unstructured==0.15.13
unstructured-client==0.25.9
urllib3==2.2.3
uvicorn==0.31.0
uvloop==0.20.0
wasabi==1.1.3
watchfiles==0.24.0
weasel==0.4.1
websocket-client==1.8.0
websockets==13.1
whatthepatch==1.0.6
wrapt==1.16.0
yapf==0.40.2
yarl==1.13.1
zipp==3.20.2
